requirements.txt
opencv-python>=4.5.0
numpy>=1.19.0
python-can>=3.3.0
RPi.GPIO>=0.7.0
